src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-identity==1.13.0
azure-servicebus==7.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.0
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
